refactor: change field requirements
